@@ -1,12 +1,12 @@
-﻿using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using SFA.DAS.RoatpGateway.Web.Settings;
 
 namespace SFA.DAS.RoatpGateway.Web.StartupExtensions
 {
     public static class CacheStartupExtensions
     {
-        public static IServiceCollection AddCache(this IServiceCollection services, IWebConfiguration configuration, IHostingEnvironment environment)
+        public static IServiceCollection AddCache(this IServiceCollection services, IWebConfiguration configuration, IHostEnvironment environment)
         {
             if (environment.IsDevelopment())
             {
