@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.DataProtection;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using SFA.DAS.RoatpGateway.Web.Settings;
 using StackExchange.Redis;
 
@@ -9,7 +9,7 @@ namespace SFA.DAS.RoatpGateway.Web.StartupExtensions
 {
     public static class DataProtectionStartupExtensions
     {
-        public static IServiceCollection AddDataProtection(this IServiceCollection services, IWebConfiguration configuration, IHostingEnvironment environment)
+        public static IServiceCollection AddDataProtection(this IServiceCollection services, IWebConfiguration configuration, IHostEnvironment environment)
         {
             if (!environment.IsDevelopment())
             {
