@@ -1,11 +1,11 @@
-﻿using Microsoft.AspNetCore.Authentication;
+﻿using System.Linq;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
+using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using Microsoft.AspNetCore.Authentication.WsFederation;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-using System.Linq;
-using System.Security.Claims;
-using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 using SFA.DAS.RoatpGateway.Web.Models;
 using SFA.DAS.RoatpGateway.Web.Settings;
 
@@ -53,7 +53,7 @@ namespace SFA.DAS.RoatpGateway.Web.Controllers
         }
 
         [HttpGet]
-        public IActionResult SignOut()
+        public override SignOutResult SignOut()
         {
             var callbackUrl = Url.Action("SignedOut", "Account", values: null, protocol: Request.Scheme);
 
@@ -97,7 +97,7 @@ namespace SFA.DAS.RoatpGateway.Web.Controllers
                 UseDfESignIn = _webConfiguration.UseDfeSignIn,
                 HelpPageLink = _webConfiguration.DfESignInServiceHelpUrl
             };
-            return View("AccessDenied",model);
+            return View("AccessDenied", model);
         }
     }
 }
